requirements.txt
Django~=3.2.7
django-crispy-forms~=1.12.0
python-decouple~=3.4